trail DONE

+++ Assets/Scripts/Firebase/TrailMaking.cs
@@ -5,12 +5,12 @@ public class TrailMaking
     public string time;
     public int level;
     public float scorePercent;
-    public int accuracy;
+    public float accuracy;
     public float overallTime;
     public int numberOfMistakes;
 
     public TrailMaking(string date, string time, int level, float scorePercent,
-        int accuracy, float overallTime, int numberOfMistakes)
+        float accuracy, float overallTime, int numberOfMistakes)
     {
         this.date = date;
         this.time = time;
