+++ Assets/Scripts/Firebase/WhackAmole.cs
@@ -4,13 +4,13 @@ public class WhackAmole
     public string time;
     public int level;
     public float scorePercent;
-    public int accuracy;
+    public float accuracy;
     public float overallTime;
     public float goResponseTime;
     public float noGoResponseTime;
 
     public WhackAmole(string date, string time, int level, float scorePercent,
-        int accuracy, float overallTime, float goResponseTime, float noGoResponseTime)
+        float accuracy, float overallTime, float goResponseTime, float noGoResponseTime)
     {
         this.date = date;
         this.time = time;
